+++ kit/dapp/src/app/[locale]/(private)/portfolio/layout.tsx
@@ -1,0 +1,26 @@
+import type { Metadata } from 'next';
+import type { Locale } from 'next-intl';
+import { getTranslations } from 'next-intl/server';
+import type { PropsWithChildren } from 'react';
+
+interface LayoutProps extends PropsWithChildren {
+  params: Promise<{ locale: Locale }>;
+}
+
+export async function generateMetadata({
+  params,
+}: LayoutProps): Promise<Metadata> {
+  const { locale } = await params;
+  const t = await getTranslations({
+    locale,
+    namespace: 'layout.navigation',
+  });
+
+  return {
+    title: t('portfolio'),
+  };
+}
+
+export default function PortfolioLayout({ children }: LayoutProps) {
+  return <>{children}</>;
+}
