+++ kit/dapp/src/components/blocks/my-assets-table/my-assets-table-columns.tsx
@@ -23,13 +23,10 @@ export function columns() {
       header: t('symbol-header'),
       enableColumnFilter: false,
     }),
-    columnHelper.accessor(
-      (row) => <ColumnAssetType assettype={row.asset.type} />,
-      {
-        id: t('type-header'),
-        header: t('type-header'),
-      }
-    ),
+    columnHelper.accessor('asset.type', {
+      header: t('type-header'),
+      cell: ({ getValue }) => <ColumnAssetType assettype={getValue()} />,
+    }),
     columnHelper.accessor('value', {
       header: t('balance-header'),
       meta: {
