+++ kit/dapp/src/lib/config/metadata.ts
@@ -55,11 +55,6 @@ export const metadata = {
   icons: {
     icon: '/favicon.ico',
   },
-  viewport: {
-    width: 'device-width',
-    initialScale: 1,
-    maximumScale: 1,
-  },
   robots: {
     index: true,
     follow: true,
